+++ components/gfx/GfxCard.tsx
