+++ components/layout/MainContent.tsx
