+++ routes/about.tsx
@@ -1,10 +1,10 @@
 import Layout from "@/components/layout/Layout.tsx";
 import { Handlers, PageProps } from "$fresh/server.ts";
-import Markdown from "@/components/pageComponent/Markdown.tsx";
+import Markdown from "@/components/article/Markdown.tsx";
 import { DataPath } from "@/deps/paths.ts";
 import { join } from "$std/path/mod.ts";
-import Contact from "@/components/pageComponent/Contact.tsx";
-import Article from "@/components/pageComponent/Article.tsx";
+import Contact from "@/components/layout/Contact.tsx";
+import Article from "@/components/article/Article.tsx";
 
 export const handler: Handlers<string> = {
   async GET(_req, ctx) {
