+++ routes/articles/index.tsx
@@ -1,9 +1,9 @@
 import { Handlers, PageProps } from "$fresh/server.ts";
 import { getPosts, Post } from "@/deps/posts.ts";
 import Layout from "@/components/layout/Layout.tsx";
-import PostCard from "@/components/pageComponent/PostCard.tsx";
-import Article from "@/components/pageComponent/Article.tsx";
-import ArticleBlurb from "@/components/pageComponent/ArticleBlurb.tsx";
+import PostCard from "@/components/article/PostCard.tsx";
+import Article from "@/components/article/Article.tsx";
+import ArticleBlurb from "@/components/article/ArticleBlurb.tsx";
 
 export const handler: Handlers<Post[]> = {
   async GET(_req, ctx) {
